pyproject.toml
[project]
name = "pdf-data-extractor"
version = "0.1.0"
description = "PDF data extraction webapp using Landing AI APIs"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]
